chore(WeebCentral): combine files into network.ts
