+++ src/WeebCentral/pbconfig.ts
@@ -3,7 +3,7 @@ import { ContentRating, SourceIntents, type ExtensionInfo } from "@paperback/typ
 export default {
   name: "Weeb Central",
   description: "Extension that pulls content from weebcentral.com.",
-  version: "1.0.0-alpha.9",
+  version: "1.0.0-alpha.10",
   icon: "icon.png",
   language: "en",
   contentRating: ContentRating.EVERYONE,
